Have ssl mode OPTIONAL to optionally allow for ssl
Current mode ALLOW will try to load ssl cert file at start
and exit with failure if can't find that file.
This mode tries to load cert file, and if it does not find it
it continues starting up the db.
This mode is useful for tests when run with SKIPSSL=1 option.