don't use deprecated RAND_pseudo_bytes
RAND_pseudo_bytes() is deprecated in OpenSSL 1.1.0.  This makes
compilation fail on my Arch Linux box.  The alternative to
RAND_pseudo_bytes() is RAND_bytes().

The default random implementation is using the same codepath in
OpenSSL (I only checked the source for 1.0.2k), except RAND_bytes()
doesn't set the error code.  Instead of returning an error when
RAND_bytes() returns 0, we could also just ignore the error.  This would
however make us rely on undocumented behaviour that might be broken in
the future.

Instead use RAND_bytes() and check the error code properly, to ensure
compatibility with newer OpenSSL versions.  That does mean that we
introduce an additional point of failure in the function if there is not
enough entropy available, but RAND_bytes() is used in other places, and
in the normal case there should be enough entropy available to get 8
random bytes.  The alternative would be to use RAND_bytes() and ignore
the error code, but that might break with future OpenSSL versions.