Smoketest (#68)
This tests that building and installing .deb and .rpm packages still works. Also tests that after installing the packages or running make install a database can be brought up and a simple query works. Would be good if roborivers can run this before running other tests - not much point testing if can't have an installable image. Requires Docker, which is a pretty heavy requirement, but good for roborivers.