Fix selectv when it is not first token
E.g. in CTE, UNION, etc

Sniffing first token is not reliable. We do keep track in
'has_recording' flag, but it's set during cursor open and is too late.
We need to know during beginTrans whether transaction is required for
selectv or not. This makes it so.

Similarly, stop looking at first token in SP to see if read/write. Stmt
is already parsed, just look at that result.