passing replicant_can_retry to master, not working
passing replicant_can_retry from replicat to master in
osql_send_commit_logic(). But it seems that since cnonce
is passed in, something else is breaking:
Initially we thought the issue was saving cnonce in blkseq
since replicant retrying would find the outcome of the first
try (and ireq is new every time). But that does not seem to be
what the issue was.