Have logic to check for parent/child in fork history
Maintain a bit array to record as fork history whether
process is parent or child, and use that to determine
whether current process is entitled to the db handle.

If parent keeps ownership, only process with bitmap all ones
is entitled to the db handle. If child keeps ownership,
only process with bitmap all zeros keeps ownership.