Only check git version from basic test
It is not useful to check from setup because if you commit a change
the git version will change, however the executable was built with
a previous git hash, so setup will now fail because of mismatch
between git version in the executable, and the version in current dir.
Thus, i am moving this to a check in the basic.test so we can
detect the mismatch but only in one test.

To try and catch wrong executable, we would need to take md5sum of
executable at the beginning of the tests, and check with every
running setup that the md5sum matches.