Removes the version number from the getDatabaseProductName function.
This is a simple fix to workaround the problem of there being a side effect
while calling some functions on metadata while enumerating a ResultSet returned
from other metadata functions. In this case calling getDatabaseProductName was
calling getDatabaseProductVersion, which was in turn executing a query on the
database. This call was resetting the Statement member field used to enumerate
the original getTables function on the same class.

This fix removes the call to getDatabaseProductVersion (and therefore removes
the version number in the returned product name). A longer term fix
should ensure that any call into Comdb2DatabseMetaData should not cause any side
effects to other functions. This fix will require some level of support for
multiple executing Statements on the same connection.