Invalidate cursor pgno before releasing the pagelock (#341)
* Invalidate cursor pgno before releasing the pagelock

The bug was introduced in my new snapshot code in 2015. The original idea was to keep the last visited page number even though bam_search returns not found. Instead of keeping the page number in cp->pgno, we could use dbc->lastpage for the same purpose. Since we no longer hold page lock on the last page, the cursor itself should not be pointing to the page (i.e. I should have not touched cp->pgno).

This should fix the scindex.test crashes.

* enable siasofupdate.test