Jepsen: stateful conn wrapper for register test
There's an issue with jdbc conn reuse that could allow a crashed
operation's transactional scope to be recovered by a subsequent
invocation. We now close the JDBC conn whenever throwing an error, and
retry with a fresh connection.

We also politely recover from transactions where a connection fails in
the initial setup phase of the txn, and introduce a five-second delay to
avoid spamming down nodes with requests.