Jepsen: generalize register test to multiple keys
In general, this lets us test more things faster for indefinite time
periods, without crashing. This allows us to drop the custom register
model and simplify the generators: we use jepsen.independent now to
generalize those for us.  Also fixed a client bug that wrote to all rows
instead of just the given one. Also fixed a bug in the generator which
made the test take longer than it needed to, and only emit 10 seconds of
operations instead of ops over the whole test.