Jepsen: extensive refactoring of all tests
Focused primarily on making the register test more rigorous, but also
refactored the bank, set, and dirty-reads tests as well. Things are...
still a bit of a mess but mostly functional. There's a lagging issue in
the register test causing it to hang at the end of the run, which I
haven't tracked down yet.