SQLite: Fix broken parser.
SQLite parser is broken when compiled with GCC 5.4.0 on Ubuntu 16.04. For instance:

```
$ gcc --version
gcc (Ubuntu 5.4.0-6ubuntu1~16.04.5) 5.4.0 20160609
Copyright (C) 2015 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.

$ cdb2sql mydb 'insert into t values(1)'
[insert into t values(1)] failed with rc -3 near "into": syntax error
```

It can be fixed by moving the `RENAME` rule to the end of the grammar file (See https://github.com/bloomberg/comdb2/pull/237).