fix check of return code from pmux
currently clustered tests are not working because copycomdb2 which
relies on comdb2ar fails--failure is from return code from portmux:
ar requests rte and portmux returns -1 which means failure but
ar does not treat it as such but rather continues operating on the
fd which is connection to portmux and requests to hold logs on that
fd fail. By checking for return code, ar will proceed correctly
(-1 would mean can't forward fd thus the db is most likely down).