fix issue with lingering processes from first time running pmux (#602)
Currently when we run a test suite and pmux is not running but needs to be started, setup script will start pmux and runtestcase will prepend the time to the output of setup. Because of the way it does this prepending of time (to allow for rc of setup to propagate back to runtestcase), there are few processes left lingering (runtestcase, awk and tee) after the test is completed waiting for stdout of pmux and will linger until pmux is terminated.
This patch changes the way the error propagates back to runtestcase thus fixing the issue with the lingering processes.