find what is running on port on which db is supposed to run (#762)
We are seeing some failures with 'FAILED TO BIND to port...' and this
checkin attempts to find out what is running on the port on which the db
is supposed to run. Previous attempt at this was targeted at single node
database; this checkin attempts to find out used ports for any node in
a cluster.
The culprit that was releasing port before fully exiting is qcopy test which ifixed here as well.