Add genid sorting, send qblobs after record
This checkin makes possible to get the order of the blockseq operations
in the btree according to tbl, stripe, genid, wseq.
We now send qblobs after the record because we need to know whether it is
an insrec or updrec, and in the case of insrec we need to generate a fake
genid (for ordering purpose when inserting into the temp btree), and get
the stripe where this record will go into.
wseq ensures that blobs will be read before the insrec when processing
the temp tbl.