dont get a new genid for update because
we don't gain anything if we get a new genid for updates because
if an update indeed were no in place, it means that we would need
to get two locks, one on the delete and one on the add, so inserting
the upd at the end of the table does not get us deadlock free situation.
What would be needed is to break down the upd in the add and del early,
however we would need to fetch all the blobs and all the rest of the
work that is done in upd_record(). One potential alternative is to
do the del in the upd_rec, and instead of doing the add immediately
we delay it (ex insert at the end of the tbl with all the blobs needed).