Fix querytimeout assignment in Comdb2Statement (#906)
The original code was assigning a variable to itself (and ignoring an argument to the constructor).

Note that this fix will affect behavior in existing clients to the API.  Previously, the code would actually use `-1` (no timeout) regardless of specified timeout. After the fix, it will use the specified query timeout.