cdb2api event handling
An event is an action taken by the API, such as when the API connects to the server, sends a query to the server, or reads back the results from the server.

The pull request provides an interface to register an event with a callback routine. The callback is invoked when the event is fired. This allows an application to be notified of and respond to API actions, or even modify API behavior if desired.

The pull request also implements dynamic library loading in cdb2api. Dynamic library loading and event handling combined will give the API a lot of extensibility.