cdb2api: Fast and unique cnonce generation.
The new cnonce is composed of the machine hostid, the process pid, the handle address, current epoch in microseconds and a sequence number. The epoch time and the sequence number are combined into a 64-bit integer: 52 bits for the epoch time and 12 bits for the sequence number. The allocation allows a handle to run at a maximum transaction rate of 4095 txn/us (~4 billion transactions per second) till September 17, 2112. An error will be returned if the transaction rate is higher than 4095 txn/us.

The new cnonce generation is 30% faster than the current implementation on an EC2 t2.medium instance, and guarantees uniqueness.

The pull request also simplifies cancelcnonce logic.