cdb2api: Keep a string duplicate of the last query. (#1253)
Currently we latch the query in case of an HASQL retry by remembering the address of the query pointer.
This is not safe if the content in the address changes. For instance, the following code will result in
an invalid read of freed memory on the HASQL retry.

```c
/* pseudo code */
char *sql = malloc(...);
sprintf(sql, "SELECT * FROM TABLE");
cdb2_run_statement(sql);
free(sql);
while (cdb2_next_record() == CDB2_OK); /* server crashes here. */
```