cdb2api: Unique and fast cnonce generation. (#1263)
The new cnonce is composed of the machine hostid, the process pid, the handle address, current epoch in microseconds and a sequence number. The epoch and the sequence number are combined into a 64-bit integer: 52 bits for the epoch and 12 bits for the sequence number. The allocation allows a handle to run at a maximum transaction rate of 4096 transactions per microsecond (~4 billion transactions per second) till September 17, 2112. An error will be returned if the transaction rate is higher than 4096 transactions/us.

The new cnonce generation is 370% faster than the current implementation on a BTDA machine, and most importantly, it guarantees uniqueness.

The pull request also simplifies cancelcnonce logic.