reorder table sockops to not have deadlocks on master (#901)
Reorder table sockops to not have deadlocks on master, resulting in less work done by the system (undoing aborted transaction).

When replicant node sends operations to master, they are in the order which client issued them, for instance:

begin
 insert into t1..
 insert into t2..
 insert into t1..
commit
Previously we processed them in same order on the master node. In this checkin, we now reorder these operations to look like this:

insert into t1..
insert into t1..
insert into t2..
The reordering is done by changing the key of the tmptbl [where we store the osql operations]: before the key was just a sequence number, now it is table, stripe, genid, is_record/blob, sequence. Since ADD operations dont have a genid preassigned, we will now insert ADD ops into a separate tmptbl using same key, and in process_this_session we will process UPD/DEL operations first (for a given tbl/stripe, then the ADD ops). This ensures that write operations will not get page locks in data tables in a manner resulting in a deadlock. [We still can deadlock for index access since that one is not ordered yet.]

As part of this checkin, the order of operations sent from replicant needs to change--we send record before blobs and partial index (because we need a genid to order the latter correctly), however the order which the operations is saved in the table is as before (blobs before record), so that toblock processing has not been modified.

Feature can be turned on or off live via tunable reorder_socksql_no_deadlock. When switch is on, replicant node will send in the first packet is_reorder_on to master node, and will send the osql_record before osql_blobs. Master node will save is_reorder_on flag and store that session by filling key appropriately. If is_reorder_flag is not set, the key is 0 except for seq, and thus ops will not be reordered.