Live SC Delete: Don't error out for null blobs (#1295)
* Live SC Delete: Don't error out for null blobs

Currently, blobs rebuilds imply data rebuild at the current implementation. The only case when only blob files are touched is adding a blob/vutf8 field. And since new blob/vutf8 fields are guaranteed to be NULL, ideally we should not need to worry about deleting blobs in live_sc_post_delete() in this case. (i.e. The whole "else" case in the code can be removed.) However, I am keeping the code there in case we support real blob-only rebuild in the future as ignoring NULL blobs is more robust.

* add more comments