Always fail a transaction which inserts and upd/del same record via cascade (#1425)
We currently have unpredictable behavior when inserting and updating or deleting the inserted record within the same transaction. The reason is that fndlen returned from loading the record can be 0 or dtasize depending on whether stripe has any records. This PR fixes that by always failing such transactions. Also adds a testcase.