Heuristic Buffer Pool Trickle
Currently we keep trickling the buffer pool to maintain above a certain
clean page percentage. However the approach fails to archieve a good
balance between different workloads: for a read-intensive workload, one
would want to keep the clean page percentage rather high to avoid dirty
page eviction on a page read; for an update-intensive workload especially
when the entire dataset fits in the cache, one would want to keep the
clean page percentage rather low to avoid unnecessary page writes.

The pull request implements a heuristic buffer pool trickling mechanism.
Instead of flushing dirty pages to a fixed percentage, estimate how many
pages are needed in the near future from the buffer pool statistics, and
pre-flush that many pages, such that the avaiable clean pages match the
the demands as closely as possible.

The pull request also changes cnonce values from the same handle to be
monotonically increasing. This greatly increases the fill factor of blkseq
and decreases the cache misses in blkseq cache.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>