Temparray
A temparray is a lightweight replacement of a temptable. It is merely a sorted array.
If the number of elements is greater than a threshold (512 by default),
or the in-memory data size exceeds a pre-configured cache size (512KiB by default),
a temparray will fall back to a temptable.

A temparray is more efficient than a temptable, thanks to its simple data structure.
Besides, it uses far less memory than a temptable to process small and medium-sized requests.