Add ability to whitelist databases allowed to connect to a given db (#1472)
Add ability to whitelist databases allowed to connect to a given db.
Db list is passed in lrl in the form
 remsql_whitelist databases d1 d2 d3.
If an other DB makes a connection to this DB, we will only continue processing
request if that DB's name is in the whitelist, otherwise it will receive an
error.
Also added ability to send message to db to add/delete/dump to whitelist content.