Debug berkeley lockers upon lock/unlock (#1484)
If DEBUG_LOCKS is defined will print info and stactrace everytime a berkeley lock is locked and everytime is released.
Note: enabling this will spew quite a bit (when lots of lock activity) and the chance to see deadlocks is higher since we now hold on to locks for longer while we print this info.

* Show all the children lockers involved in deadlock
* Have lockput() show correct file and page being unlocked (just like lockget())
* Print all active objects to see full deadlock cycle

Instead of dumping active lockers, which will skip lockers that
have only one lock in WAIT status, to completely show all deadlock
cycles we need to dump all active objects instead. Note that we
loose the nice sorting by thread, but that should be easy to
do manually.