Add timing accounting: accumulate time spent in a given function (#1537)
Add timing accounting: accumulate time spent in a given function:
Make it simple to get timing information how much time was spent in a given function, for debug purposes. Uses a global array to accumulate timing for a given function/feature. As an example, in this PR we measure time to add record, add index, and add to temp table. Adding new timing info is trivial by adding one more element (ex NEWNAME) to the array, and calling:

ACCUMULATE_TIMING(NEWNAME, 
CODE_TO_TIME...;
);

This is enabled only for debug mode: in the non-debug build the functions are #defined to {} so it consumes no cpu time at all.