ix_find to get write lock in updrec and delrec (#1589)
In upd_rec_indices and del_rec_indices we call ix find in readl lock first to verify that records exists, then subsequently get write lock to update the record. This is the cause for a deadlocks and instead, by getting the record in write lock from the beginning, we turn the deadlocks into waits, thus improving performance by 20% in some sysbench tests, but more importantly by not busying the cpu undoing work in the case of a deadlock (which aborts one of the transactions).
This was part of #1345 but checking it in separately because it is relevant on its own.