Having sql_tick always recover deadlock when bdb lock is desired
The function sql_tick() does not call recover_deadlock() for 'SELECT SLEEP(N)',
'ANALYZE' and queries on temptables. If a node where such a query gets
processed is the master and being downgraded, or the node is a replicant and
being upgraded, the downgrade or upgrade will be blocking on the bdb write lock
till the query is finished.

The patch changes sql_tick() to always recover deadlock when bdb lock is desired.

(DRQS 144712120)