Faster Analyze (#1652)
Significantly speed up analyze (2.5x ~ 3x) by reducing the cost of sorting.
Instead of sampling individual entries and sorting them in a temptable,
we sample and sort pages. This way we leverage the fact that entries
on the same page are already sorted and are able to greatly reduce sorting cost.