cdb2api: Renaming (un)install_static_libs
Rename `install_static_libs` and `uninstall_static_libs` to `enable_static_libs` and `disable_static_libs` respectively. We've discovered a bug in the hungserv plugin. Therefore we need a way to disable the feature in buggy versions but still enable it in later versions. Renaming the configuration names is the simplest solution.