Fix SQL mspace race condition in DOHSQL
An SQlite engine's mspace isn't thread-safe because it's supposed to be accessed by the engine only. DOHSQL breaks the assumption:
While the master engine is casting a Mem object which needs to reallocate zMalloc from its mspace, a child engine may decide to reuse or discard a cached Mem object and therefore free zMalloc in the Mem object which is also allocated from the master's mspace. They can race with each other.

An obvious fix to this is to make an SQL mspace thread-safe. However this would incur unnecessary locking overhead for non-parallelizable queries. Instead we use a mutex to guard the Mem objects. We just need to ensure that we always hold the mutex whenever casting, reusing or discarding a Mem object inside the DOHSQL subsystem.

(DRQS 150091313)