Verify: stop verify when client disconnected and not verbose (#1992)
Verify: stop verifying if client disconnected (and verbose is turned off).
The bug is that if we don't specify verbose in the verify command, verify continues to process even after cdb2sql client has disconnected. The reason is because we currently have no way to check whether client is still connected if verbose is off. With this checkin we now pass in sbuf object which we can perform such check. Now we have to avoid printing directly messages to sbuf because that breaks clients connecting via cdb2sql.
Also in this checkin:

*   Added to insert_lots.test checks to ensure all threads in verify thdpool exited when client exits
*   Moved sequential verify to be performed by the thdreadpool just like the parallel one