Use Canonical Name everywhere (#2173)
If the master node has a Canonical Name different from its local hostname,
and its Canonical Name is written in `hostname.lrl`, it'll think that the
master is down and then call for an election, repeatedly.

This is because `gbl_myhostname`, which is the local hostname, is used in
the `net/event` layer to check for localhost, whereas `gbl_node`, which is
the host's Canonical Name, is used to create a `net`. If `gbl_myhostname`
and `gbl_node` are different, the `net/event` layer will fail to filter
out net messages from localhost to localhost.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>