Use thread local variable in get_area()
End result is same underlying behavior but without the penalty
of pthread_getspecific() which allows some savings of about 1%
in a simple select test case (as reported by valgrind).

The usage here is a bit peculiar in having redundant pointers
to the same zone because there is no way to cleanup __thread
variables in c. We use the pthread_key variable to perform
cleanup on thread exit, and for checking whether zone is null
(which is done hundreds of thousand of times) we use the
__thread variable.

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>