Make tests work for pmux running on non default port #2275
Testsuite has ability to pass PMUXPORT=xxxxx as parameter to make so we can run pmux
at any port, not just the default one. Many of the tests do not use the $CDB2_OPTIONS variable
so they don't work with pmux running on the non default pmux.
This patch fixes many of the tests in question.
Also small changes to the cluster script and testrunner.

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>