cdb2api bug fixes: pmux, garbage config and the local tier (#2367)
The change fixes 3 bugs in cdb2api.

1. Unnecessary pmux lookup in DIRECT-CPU mode
In DIRECT-CPU mode the API always looks up the database port. This costs an
unnecessary round trip to pmux when there is already a cached connection in
the sockpool. We should do so only when the API needs to make a connection on
its own.

2. Garbage configuration file
The API attempts to process "(null).cfg". This is caused by passing a NULL
database name to snprintf when formatting the database configuration file path.

3. The "local" tier does not mean localhost
The API performs dbinfo query and pulls all nodes in the cluster regardless of tier.
However dbinfo should be skipped when "local" is specified as tier.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>