DBA user
This PR introduces the concept of DBA user. It's a user that
gets created automatically (`create_dba_user`) on server start
if it does not exist already. The user has OP privileges and
has an extra bit of protection in that its properties cannot
be changed with `lock_dba_user` tunable set.

* The user is created only when authentication is not enabled.
* DBA user name is hardcoded to 'dba' with an empty password.
* Create this user only on master
* Disable lock_dba_user by default

Signed-off-by: Nirbhay Choubey <nchoubey@bloomberg.net>