Eventlog: fingerprint and cnonce for begin/commit/rollback
We need accurate fingerprint and cnonce for begin/commit/rollback
sql entries in order to differentiate statements in a transaction
and commit them correctly and in order.

Changed cdb2_sqlreplay to group sql stmts by cnonce into transactions
and replay them only once we see the commit/rollback entry,
such that we get close to the order in which they were commited originally.
(To emulate the original order/timing we'd need to use the 'txn' entry,
and run each txn stmt in parallel in its own thread).

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>