Tweak systbl_proc_leak test (#2501)
The test works only when there is one single SQL thread in the database.
However because of the extra cleanup work after a statement is finished,
an SQL thread isn't always immediately available for the next statement
and hence a new SQL thread is spawned. It causes the test to fail intermittently.

The change makes sure that the test database can have 1 SQL thread at most.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>