Use buffer to unpack protobuf query messages (#2526)
Use a small buffer as part of clnt structure to unpack protobuf query.
This is similar to 63c9364ed5 (#1645) which added buffer to cdb2api.

As a measure of savings, running basic.test under valgrind with and without
this patch shows a reduction of 136K calls to comdb2_malloc_static()
(via the protobuf_message_unpack() callpath) for the test's 5146 sql queries
(out of 1316K total mallocs, so almost 10% less mallocs).

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>