Copy blobs only once (#2565)
When retrieving a blob, we copy it from berkdb (if uncompressed) or bdb
(if compressed) to sqlite, so that the memory is owned by sqlite. However,
we can get rid of the extra copy by passing sqlite's allocator to berkdb and
bdb and allocate memory directly from it. This way it'll be faster and require
less memory.

(DRQS 163035031)

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>