Move logfiles into a backup dir if feature enabled (#2583)
Move the logfiles instead of deleting them if lrl option is set:
```
   location backup_logfiles_dir /bigstore/backup/%dbname
```
This will allow for ability to restore to any point in time given that
logs will always be available. This can work well with a reasonable
maintenance schedule to backup to tape and cleanup log files
older than a preset value (ex a month).

If directory ends with `%dbname`, we will substitute the actual dbname
to use as directory where the logfiles will be backed up.

Note that if the backup_logfiles_dir is an nfs mounted directory
which is accessible from all the nodes in a cluster, correctness
of the backedup logfiles is ensured because only the master is able
to move logfiles to the backup directory (amongst other restrictions,
we only delete/move a logfile if there is no reference to a transaction
in blkseq).

Added clustered test which exercises this feature.

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>