Looks like it works still after merge, but it seems is now slower
This appears now twice as slow as master but only on 32 core machines.
On 4 core machines it appears equivalent (to the temparray)
regardless of the transaction size.
I think because they were performing equal when i was testing the pull request
and also the reason I left this without merging because I was not
sure about the speedup.
But now it is missbehaving--something is not right so investigating.