Synchronous osql cancellation
An osql cancellation is done asynchronously in the sense that a replicant
doesn't wait for it to complete on the master node. If there's a large number
of rollbacks or client disconnects, and the master can't cancel them faster
than they arrive, these requests will pile up on the master.

This patch makes osql cancellation synchronous, so that the master can be
protected in this case.

(DRQS 164512050)

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>