Speedup truncate: cache version_num in process_unused_files (#2660)
When a db has many tables tables, truncate will take a long time because
delfiles will go through the entire directory to find files matching pattern
to be put in list for deletion. The inneficiency is because we fetch
version_num for every file in the directory: this checkin improves that
by caching those values in an array.
Also dont check if table is under schemachange when we have tbllock on it.

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>