Add missing osql type to be consistent with r7 (#2666)
When refactoring sqltypes, ony type was missed because was not in use
in master branch, however it is in r7 and thus needs to be added to
masterbranch as well so enum for messages will match with r7.
Also make enum to be the value we specified in the macro.

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>