Fix issues with inserts and adding first index
When we are running a schemachange adding the first index,
in the very brief period of time after sc_pointer has been set to 0xffff
and until getting tbl lock in write mode will be missing inserts to the
index btrees. This checkin fixes that by adding to the ops temptbl when
adding such first index.

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>