Only get tablelock once for repeated USEDB of same tbl (#2742)
Currently we get tablelock for every OSQL_USEDB message. In this case:
```
begin
insert into t1
insert into t2
insert into t1
commit
```
the inserts get processed in order t1,t1,t2 is because of osql table reordering.
This change will skip getting tablelock for subsequent USEDBs of the same table
if we already got it once before.

We set iq->usedb to `_comdb2_static_table` upon starting a transaction,
so we are guaranteed to always get the tablelock for the first USEDB,
still I added asserts to verify that we indeed have the tablelock (in debug build).

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>