Fix implicit conversion from String to Datetime (#2747)
To compare 2 operands of different types, one of the operands is first promoted
to the type of the other operand. After the comparison, the changes to the flags
on the promoted operand are undone, demoting the operand to its original type.
However, there're a few issues in this logic, when it comes to comparing
a String with a Datetime.

1) Invalid datetime strings causing arbitrary results

Consider the following table and query

```
Table t

| col (datetime)|
|---------------|
|   2021-01-01  |
|   2021-01-01  |

SELECT * FROM t WHERE col < "ABC"
```

The literal "ABC" isn't a valid datetime. Therefore when converting it
to Datetime, `sqlite3VdbeMemDatetimefy()` fails and returns an error.
Unfortunately `sqlite3MemCompare()` ignores the error, calls `dttz_cmp()`,
and ends up comparing an uninitialized `dttz_t` field in the Mem object
of "ABC" with the Datetime column. The results are hence undefined.

To fix this, we attempt to compare the 2 values as Datetime values,
and fall back to strcmp if it fails.

2) Valid datetime strings working by coincidence

Consider the following table and query

```
Table t

| col (datetime)|
|---------------|
|   2021-01-01  |
|   2021-01-01  |

SELECT * FROM t WHERE col < '2022-01-01'
```

When comparing the String operand with the 1st Datetime value in the table,
the operand is promoted to Datetime. `sqlite3VdbeMemDatetimefy()`, however,
frees `Mem.z` and sets it to null. After the comparison, the operand's flags
are restored to `MEM_Str`. At this point, we have a `MEM_Str` operand, with
a null string value, and a `dttz_t` field of January 1st, 2022.

Now the operand is compared to the 2nd value. `sqlite3VdbeMemDatetimefy()`
is going to fail, as the operand is not a legitimate string any more.
`sqlite3MemCompare()`, as we mentioned, conveniently ignores the conversion
error, compares the previously converted `dttz_t` field to the 2nd row value,
and returns a correct result.

The logic here is obviously flawed. The patch improves it by keeping an operand's
flags if it's converted from a String to a non-String type, and has lost its string
value during the conversion. So in the case above, the String literal will be converted
to a Datetime exactly once, and stays a Datetime for the rest comparisons.

3) Datetime strings without timezone resulting in undefined row orders in order-by

The function `str_to_dttz()` expects a valid timezone. However, in an order-by query,
MEM_Str objects deserialized from the VDBESorter do not have their `tz` set, causing
`str_to_dttz()` to return an error. Again, because `compareDateTimeInterval()` does not
check for any type error, it may cause the VDBESorter to return an undefined row order.

To fix it, we simply use the `tzname` from the `clnt`, if `tz` isn't present.

(DRQS 165247061)

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>