Add hardware crc32 function calls for arm arch
Use family of functions __crc32cb() and friends to compute crc32c in hardware
if such feature is present in the cpu.

Simple test is attached and shows a speedup of about 6 times on the `armv7l` architechture:
```
    software:  1318717usec
    hardware:   227700usec
```

and a speedup of 10 times for the same test on the `aarch64` architechture:

```
    software:   284895usec
    hardware:    27514usec
```

for completeness, the intel architechture gives a speedup of > 10 times:

```
    software:   329866usec
    hardware:    26930usec
```

NB: The above results for the test were done on different types of processors
(ie. not server-grade) and as such it is not appropriate to compare the runtimes
between architechtures.

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>