Recreate tool hardlinks during installation
When the 'comdb2' executable is installed, CMake can fix up its RPath to
what its final value should be. Additionally, using install() in
PROGRAMS mode copies the source to the destination, which breaks the
hardlink relationship between the tools and 'comdb2'.

Instead, we create hardlinks to 'comdb2' has been built but make a new
set of hardlinks to 'comdb2' after it has been installed.

Signed-off-by: Mike Seplowitz <mseplowitz@bloomberg.net>