Correctly handle unix time of -1 (#2873)
db_struct2time() returns -1 on error, as well as for "1969-12-31T23:59:59".
There's no way to tell the two cases apart and we wind up treating "1969-12-31T23:59:59" as an error.

To fix it, errno is set if there's an actual failure. When db_struct2time() returns -1, if errno is 0,
it's "1969-12-31T23:59:59"; otherwise it's an error.

(DRQS 166367318)

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>