Reset uuid after a write commits (#2880)
We currently log in eventlog select stmts with a uuid if it follows an insert
because we dont clean the clnt->osql.uuid variable. This checkin does just that:
resets it to zero after the write has commited. Also added a testcase to verify
that the select does not get logged with a uuid value.

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>