Serializing protobuf to evbuffer
The patch is a re-implementation of #2184 for the new libevent-appsock logic.
It uses significantly less memory and runs faster than the naive approach where
an entire protobuf structure is serialized to memory, which then is written to
the underlying file descriptor.

Additionally it accounts for protobuf memory in libevent-appsock by using
the `pb_alloc' allocator.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>